requirements.txt
pillow==6.1.0
networkx==2.1
torch==1.3.1
numpy==1.16.3
scipy==1.3.0rc1
torchvision==0.2.2.post3
PyYAML==3.13
